% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanvector.R
\name{assert_class}
\alias{assert_class}
\title{Checks whether a vector is from a specified class.}
\usage{
assert_class(vector, required_class)
}
\arguments{
\item{vector}{The object to check.}

\item{required_class}{character. A string specifying the class the object should inherit from.}
}
\value{
logical. Returns \code{TRUE} if the object is of the required class; otherwise, prints a warning and returns \code{FALSE}.
}
\description{
Assert that an object inherits from a specified class.
}
\examples{
# Example: Passes and returns TRUE because 1:3 is of class "integer"
assert_class(1:3, "integer")

# Example: Prints warning and returns FALSE because 1:3 is not of the
# class "character"
assert_class(1:3, "character")

}
