% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanvector.R
\name{convert_class}
\alias{convert_class}
\title{Convert an object to a different class.}
\usage{
convert_class(vector, required_class, custom_transform_function = NULL)
}
\arguments{
\item{vector}{An object to be coerced.}

\item{required_class}{character. A string specifying the target class.}

\item{custom_transform_function}{A custom function to transform the vector to the required class. Default is \code{NULL}.}
}
\value{
vector. coerced to the required class if necessary.
}
\description{
Convert an object to a specific built-in class if it is not already of that class.
If a class is not built-in and no custom function is provided, an error will be thrown.
Convert an object to a specified class outside of the built-in types using a custom function.
If the conversion fails and the resulting vector is not of the desired class, and error will be thrown.
}
\examples{
# Example: Coerce numeric to character
convert_class(1:3, "character")

# Example: No coercion needed (already integer)
convert_class(1:3, "integer")

# Example: Attempt to coerce character to numeric
convert_class(c("1", "2", "3"), "numeric")

# Example: Attempt to coerce numerical vector into list.
convert_class(1:3, "list", as.list)
}
